Add files via upload
While the Think or Swim stock platform is one of the best out there in many ways, many people would rather be able to watch their stocks using Excel. Excel is a tool they are familiar with, and with the right code in the background, can offer more data in one place than any of the stock listings available on the Think or Swim platform.